requirements.txt
yfinance
notion-client==2.2.1
python-dotenv
httpx[socks]
socksio
flask
schedule
